iox-#2330 Refactor call to `iox_getenv_s` for type safety
Cast `info.total_size` to unsigned int in the `iox_getenv_s` call to ensure type matching and eliminate potential issues with type mismatches. This change improves code robustness and ensures adherence to function signature requirements.

+++ iceoryx_posh/source/roudi/roudi.cpp
@@ -607,7 +607,8 @@ std::string ServiceManagementSystemd::getEnvironmentVariable(const char* const e
 
     str.unsafe_raw_access([&](auto* buffer, auto const info) {
         size_t actualSizeWithNull{0};
-        auto result = IOX_POSIX_CALL(iox_getenv_s)(&actualSizeWithNull, buffer, info.total_size, env_var)
+        auto result = IOX_POSIX_CALL(iox_getenv_s)(
+                          &actualSizeWithNull, buffer, static_cast<unsigned int>(info.total_size), env_var)
                           .failureReturnValue(-1)
                           .evaluate();
